refactir(dashboard): move /support into (organization-active)
